requirements.txt
aniso8601==3.0.2
Django==2.1.5
django-ckeditor==5.6.1
django-cors-headers==2.4.0
django-filter==2.1.0
django-graphql-extensions==0.0.4
django-graphql-jwt==0.2.0
django-js-asset==1.1.0
djangorestframework==3.8.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
Pillow==5.4.1
promise==2.2.1
psycopg2-binary==2.7.5
PyJWT==1.7.1
pytz==2018.9
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
